backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.3.0

# HTTP client
httpx>=0.24.0,<0.25.0

# Environment & Config
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Amadeus SDK
amadeus==8.1.0
